chore(gateway): rename cairo_lang_compiler file
